#78: refactoring
