+++ lib/util/init.js
@@ -9,6 +9,7 @@ const InitConfig = require('./init.config');
 const inquirer = require('inquirer');
 const Util = require('./util');
 const fs = require('node:fs');
+const path = require('node:path');
 
 /**
  * CLI helper class
@@ -23,7 +24,7 @@ const Init = {
      * @returns {Promise.<void>} -
      */
     async initProject(properties, credentialName) {
-        Util.checkPathForCloud();
+        Init._checkPathForCloud();
 
         const skipInteraction = Util.skipInteraction;
         if (!properties) {
@@ -195,7 +196,7 @@ const Init = {
      * @returns {Promise.<void>} -
      */
     async joinProject() {
-        Util.checkPathForCloud();
+        Init._checkPathForCloud();
 
         const responses = await inquirer.prompt([
             {
@@ -375,7 +376,7 @@ const Init = {
      * @returns {Promise.<boolean>} success flag
      */
     async upgradeProject(properties, initial, repoName) {
-        Util.checkPathForCloud();
+        Init._checkPathForCloud();
 
         let status;
         const versionBeforeUpgrade = properties?.version || '0.0.0';
@@ -411,8 +412,31 @@ const Init = {
         return true;
     },
     /**
+     * check if git repo is being saved on a cloud service and warns the user
+     *
+     * @private
+     * @returns {void} throws errors if problems were found
+     */
+    _checkPathForCloud() {
+        const absolutePath = path.resolve('');
+        // popular cloud services and their respective default name for the absolute path
+        const cloudServices = ['Dropbox', 'OneDrive', 'Google Drive'];
+        for (const variable in cloudServices) {
+            if (absolutePath.includes(cloudServices[variable])) {
+                Util.logger.warn(
+                    `It seems your project folder will be synchronized via '${cloudServices[variable]}'. This can reduce the overall performance of your computer due to conflicts with Git.`
+                );
+                Util.logger.warn(
+                    `We strongly recommend moving your project folder outside of the '${cloudServices[variable]}' folder.`
+                );
+            }
+        }
+    },
+
+    /**
      * finds credentials that are set up in config but not in auth file
      *
+     * @private
      * @param {TYPE.Mcdevrc} properties javascript object in .mcdevrc.json
      * @returns {string[]} list of credential names
      */
