+++ lib/util/util.js
@@ -4,7 +4,6 @@ const TYPE = require('../../types/mcdev.d');
 const MetadataDefinitions = require('./../MetadataTypeDefinitions');
 const packageJsonMcdev = require('../../package.json');
 const child_process = require('node:child_process');
-const path = require('node:path');
 const process = require('node:process');
 const toposort = require('toposort');
 const winston = require('winston');
@@ -453,26 +452,6 @@ const Util = {
         return str;
     },
 
-    /**
-     * check if git repo is being saved on a cloud service and warns the user
-     *
-     * @returns {void} throws errors if problems were found
-     */
-    checkPathForCloud() {
-        const absolutePath = path.resolve('');
-        // popular cloud services and their respective default name for the absolute path
-        const cloudServices = ['Dropbox', 'OneDrive', 'Google Drive'];
-        for (const variable in cloudServices) {
-            if (absolutePath.includes(cloudServices[variable])) {
-                Util.logger.warn(
-                    `It seems your project folder will be synchronized via '${cloudServices[variable]}'. This can reduce the overall performance of your computer due to conflicts with Git.`
-                );
-                Util.logger.warn(
-                    `We strongly recommend moving your project folder outside of the '${cloudServices[variable]}' folder.`
-                );
-            }
-        }
-    },
     /**
      * get key of an object based on the first matching value
      *
